requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
supabase==2.7.4
aiohttp==3.9.5
pydantic==2.5.3
tenacity==8.5.0
fastapi==0.112.0
jinja2==3.1.4
uvicorn==0.30.3

